requirements.txt
gradio
huggingface_hub
python-dotenv
keybert
transformers
langdetect
